requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
langchain>=0.1.12
langchain-openai>=0.0.8
langchain-community>=0.0.28
faiss-cpu==1.7.4
openai>=1.12.0
requests>=2.31.0 